ethabi
Struct Contract

pub struct Contract {
    pub constructor: Option<Constructor>,
    pub functions: HashMap<String, Vec<Function>>,
    pub events: HashMap<String, Vec<Event>>,
    pub receive: bool,
    pub fallback: bool,
}

API building calls to contracts ABI.

Fields
constructor: Option<Constructor>

Contract constructor.

functions: HashMap<String, Vec<Function>>

Contract functions.

events: HashMap<String, Vec<Event>>

Contract events, maps signature to event.

receive: bool

Contract has receive function.

fallback: bool

Contract has fallback function.

Implementations

impl Contract

pub fn load<T: Read>(reader: T) -> Result<Self>

Loads contract from json.

pub fn constructor(&self) -> Option<&Constructor>

Creates constructor call builder.

pub fn function(&self, name: &str) -> Result<&Function>

Get the function named name, the first if there are overloaded versions of the same function.

pub fn event(&self, name: &str) -> Result<&Event>

Get the contract event named name, the first if there are multiple.

pub fn events_by_name(&self, name: &str) -> Result<&Vec<Event>>

Get all contract events named name.

pub fn functions_by_name(&self, name: &str) -> Result<&Vec<Function>>

Get all functions named name.

pub fn functions(&self) -> Functions<'_>

Iterate over all functions of the contract in arbitrary order.

pub fn events(&self) -> Events<'_>

Iterate over all events of the contract in arbitrary order.