Crate ethabi

Ethereum ABI encoding decoding library.

Modules

param_type
Function and event param types.

token
ABI param and parsing for it.

util
Utils used by different modules.

Structs

Constructor
Contract constructor specification.

Contract
API building calls to contracts ABI.

Error
The Error type.

Event
Contract event.

EventParam
Event param specification.

Events
Contract events interator.

Function
Contract function specification.

Functions
Contract functions interator.

Log
Decoded log.

LogParam
Decoded log param.

Param
Function param.

RawLog
Ethereum log.

RawTopicFilter
Raw topic filter.

TopicFilter
Topic filter.

Enums

ErrorKind
The kind of an error.

ParamType
Function and event param types.

Token
Ethereum ABI params.

Topic
Acceptable topic possibilities.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

decode
Decodes ABI compliant vector of bytes into vector of tokens described by types param.

encode
Encodes vector of tokens into ABI compliant vector of bytes.

Type Definitions

Address
ABI address.

Bytes
ABI bytes.

FixedBytes
ABI fixed bytes.

Hash
Commonly used FixedBytes of size 32

Int
ABI signed integer.

Result
Convenient wrapper around std::Result.

Uint
ABI unsigned integer.