Crate ethabi

Ethereum ABI encoding decoding library.

Re-exports

pub use ethereum_types;

Modules

param_type

Function and event param types.

token

ABI param and parsing for it.

Structs

AbiError

Contract error specification.

Constructor

Contract constructor specification.

Contract

API building calls to contracts ABI.

Event

Contract event.

EventParam

Event param specification.

Events

Contract events iterator.

Function

Contract function specification.

Functions

Contract functions iterator.

Log

Decoded log.

LogParam

Decoded log param.

Param

Function param.

RawLog

Ethereum log.

RawTopicFilter

Raw topic filter.

TopicFilter

Topic filter.

TupleParam

Tuple params specification

Enums

Error

Ethabi errors

ParamType

Function and event param types.

StateMutability

Whether a function modifies or reads blockchain state

Token

Ethereum ABI params.

Topic

Acceptable topic possibilities.

Traits

FunctionOutputDecoder

Contract functions generated by ethabi-derive

LogFilter

Common filtering functions that are available for any event.

ParseLog

trait common to things (events) that have an associated Log type that can be parsed from a RawLog

Functions

decode

Decodes ABI compliant vector of bytes into vector of tokens described by types param.

encode

Encodes vector of tokens into ABI compliant vector of bytes.

long_signature

Returns the full Keccak-256 hash of the signature of the given params

short_signature

Returns the first four bytes of the Keccak-256 hash of the signature of the given params

Type Definitions

Address

ABI address.

Bytes

ABI bytes.

FixedBytes

ABI fixed bytes.

Hash

Commonly used FixedBytes of size 32

Int

ABI signed integer.

Result

Ethabi result type

Uint

ABI unsigned integer.

Word

ABI word.