Crate ethabi

Ethereum ABI encoding decoding library.

Re-exports

pub use crate::param_type::ParamType;

pub use crate::token::Token;

pub use ethereum_types;

Modules

param_type

Function and event param types.

token

ABI param and parsing for it.

Structs

Constructor

Contract constructor specification.

Contract

API building calls to contracts ABI.

Event

Contract event.

EventParam

Event param specification.

Events

Contract events iterator.

Function

Contract function specification.

Functions

Contract functions iterator.

Log

Decoded log.

LogParam

Decoded log param.

Param

Function param.

RawLog

Ethereum log.

RawTopicFilter

Raw topic filter.

TopicFilter

Topic filter.

TupleParam

Tuple params specification

Enums

Error

Ethabi errors

StateMutability

Whether a function modifies or reads blockchain state

Topic

Acceptable topic possibilities.

Traits

FunctionOutputDecoder

Contract functions generated by ethabi-derive

LogFilter

Common filtering functions that are available for any event.

ParseLog

trait common to things (events) that have an associated Log type that can be parsed from a RawLog

Functions

decode

Decodes ABI compliant vector of bytes into vector of tokens described by types param.

encode

Encodes vector of tokens into ABI compliant vector of bytes.

Type Aliases

Address

ABI address.

Bytes

ABI bytes.

FixedBytes

ABI fixed bytes.

Hash

Commonly used FixedBytes of size 32

Int

ABI signed integer.

Result

Ethabi result type

Uint

ABI unsigned integer.

Word

ABI word.