Struct ethabi::Contract

pub struct Contract {
    pub constructor: Option<Constructor>,
    pub functions: HashMap<String, Function>,
    pub events: HashMap<String, Event>,
    pub fallback: bool,
}

API building calls to contracts ABI.

Fields
constructor: Option<Constructor>

Contract constructor.

functions: HashMap<String, Function>

Contract functions.

events: HashMap<String, Event>

Contract events.

fallback: bool

Contract has fallback function.

Methods

impl Contract

fn load<T: Read>(reader: T) -> Result<Self>

Loads contract from json.

fn constructor(&self) -> Option<&Constructor>

Creates constructor call builder.

fn function(&self, name: &str) -> Result<&Function>

Creates function call builder.

fn event(&self, name: &str) -> Result<&Event>

Creates event decoder.

fn functions(&self) -> Functions

Iterate over all functions of the contract in arbitrary order.

fn events(&self) -> Events

Iterate over all events of the contract in arbitrary order.

fn fallback(&self) -> bool

Returns true if contract has fallback